+++ src/application/NestEggsManager.ts
@@ -3,6 +3,7 @@ import { BaseManager } from '../structures/BaseManager';
 import { Dict } from '../structures/Dict';
 import { FetchOptions, Include } from '../common';
 import { Egg } from '../common/app';
+import caseConv from '../util/caseConv';
 import endpoints from './endpoints';
 
 export class NestEggsManager extends BaseManager {
@@ -26,6 +27,29 @@ export class NestEggsManager extends BaseManager {
         this.cache = new Dict<number, Egg>();
     }
 
+    _patch(data: any) {
+        if (data?.data) {
+            const res = new Dict<number, Egg>();
+            for (let o of data.data) {
+                let e = caseConv.toCamelCase(o.attributes) as Egg;
+                e.createdAt = new Date(e.createdAt);
+                e.updatedAt &&= new Date(e.updatedAt);
+                res.set(e.id, e);
+            }
+
+            if (this.client.options.eggs.cache)
+                res.forEach((v, k) => this.cache.set(k, v));
+
+            return res;
+        }
+
+        let e = caseConv.toCamelCase(data.attributes) as Egg;
+        e.createdAt = new Date(e.createdAt);
+        e.updatedAt &&= new Date(e.updatedAt);
+        this.cache.set(e.id, e);
+        return e;
+    }
+
     adminURLFor(id: number): string {
         return `${this.client.domain}/admin/nests/egg/${id}`;
     }
@@ -40,16 +64,10 @@ export class NestEggsManager extends BaseManager {
             if (e) return Promise.resolve<any>(e);
         }
 
-        const data: any = await this.client.requests.get(
-            (id ? endpoints.nests.eggs.get(nest, id) : endpoints.nests.eggs.main(nest)),
+        const data = await this.client.requests.get(
+            id ? endpoints.nests.eggs.get(nest, id) : endpoints.nests.eggs.main(nest),
             options, this
         );
-
-        const res = new Dict<number, Egg>();
-        for (const egg of data.data) {
-            this.cache.set(egg.attributes.id, egg.attributes);
-            res.set(egg.attributes.id, egg.attributes);
-        }
-        return res as any;
+        return this._patch(data) as any;
     }
 }